Move validator to Handler class
