+++ src/Feature/Animal/Models/Commands/CreateAnimalCommand.cs
@@ -12,37 +12,4 @@ namespace Animal.Models.Commands
         public int Age { get; set; }
         public string Species { get; set; }
     }
-
-    public class Validator : AbstractValidator<CreateAnimalCommand>
-    {
-        public Validator()
-        {
-            //this.RuleFor(x => x.Pet)
-            //    .NotNull()
-            //    .NotEmpty();
-
-            //this.RuleFor(x => x.Pet.Name)
-            //    .NotNull()
-            //    .Length(2, 20);
-
-            //this.RuleFor(x => x.Pet.Age)
-            //    .GreaterThan(0);
-
-            //this.RuleFor(x => x.Pet.Species)
-            //    .NotNull()
-            //    .Length(3, 20);
-
-            this.RuleFor(x => x.Name)
-                .NotNull()
-                .Length(2, 20);
-
-            this.RuleFor(x => x.Species)
-                .NotNull()
-                .Length(2, 20);
-
-
-            this.RuleFor(x => x.Age)
-                .GreaterThan(0);
-        }
-    }
 }
